requirements.txt
pip
python-telegram-bot~=12.8
mysql-connector-python~=8.0.20